+++ src/QuickMarcEditor/QuickMarcEditorRows/MaterialCharsField/MaterialCharsField.test.js
@@ -7,18 +7,21 @@ import { render } from '@folio/jest-config-stripes/testing-library/react';
 import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import { MaterialCharsField } from './MaterialCharsField';
+import Harness from '../../../../test/jest/helpers/harness';
 
 const renderMaterialCharsField = ({ name, type }) => render(
-  <Form
-    onSubmit={() => {}}
-    mutators={arrayMutators}
-    render={() => (
-      <MaterialCharsField
-        name={name}
-        type={type}
-      />
-    )}
-  />,
+  <Harness>
+    <Form
+      onSubmit={() => {}}
+      mutators={arrayMutators}
+      render={() => (
+        <MaterialCharsField
+          name={name}
+          type={type}
+        />
+      )}
+    />
+  </Harness>,
 );
 
 describe('MaterialCharsField', () => {