+++ src/hooks/index.js
@@ -1,3 +1,4 @@
 export * from './useAuthorityLinking';
 export * from './useSubfieldNavigation';
 export * from './useValidation';
+export * from './useFocusFirstFieldWithError';