+++ src/hooks/useValidation/rules.js
@@ -1,4 +1,3 @@
-import { FormattedMessage } from 'react-intl';
 import { Link } from 'react-router-dom';
 
 import {
@@ -57,56 +56,54 @@ const RULES = {
 const COMMON_VALIDATORS = [
   {
     validator: RULES.LEADER_LENGTH,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.leader.length" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.leader.length' }),
   }, {
     validator: RULES.LEADER_EDITABLE_BYTES,
-    message: (marcType) => <FormattedMessage id={`ui-quick-marc.record.error.leader.forbiddenBytes.${marcType}`} />,
+    message: (marcType) => ({ id: `ui-quick-marc.record.error.leader.forbiddenBytes.${marcType}` }),
   }, {
     validator: RULES.LEADER_POSITIONS,
-    message: (positions, link) => (
-      <FormattedMessage
-        id="ui-quick-marc.record.error.leader.invalidPositionValue"
-        values={{
-          positions,
-          link: (
-            <Link
-              to={{
-                pathname: link,
-              }}
-              target="_blank"
-            >
-              {link}
-            </Link>
-          ),
-        }}
-      />
-    ),
+    message: (positions, link) => ({
+      id: 'ui-quick-marc.record.error.leader.invalidPositionValue',
+      values: {
+        positions,
+        link: (
+          <Link
+            to={{
+              pathname: link,
+            }}
+            target="_blank"
+          >
+            {link}
+          </Link>
+        ),
+      },
+    }),
   },
   {
     tag: '001',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.controlField.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.controlField.multiple' }),
   },
   {
     validator: RULES.TAG_LENGTH,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.tag.length" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.tag.length' }),
   },
   {
     validator: RULES.TAG_CHARACTERS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.tag.nonDigits" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.tag.nonDigits' }),
   },
   {
     validator: RULES.EMPTY_SUBFIELDS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.subfield" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.subfield' }),
   },
   {
     tag: '008',
     validator: RULES.EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.008.empty" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.008.empty' }),
   }, {
     tag: '008',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.008.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.008.multiple' }),
   },
 ];
 
@@ -115,52 +112,51 @@ const BASE_BIB_VALIDATORS = [
   {
     tag: '008',
     validator: RULES.FIXED_FIELD_POSITIONS,
-    message: (name) => <FormattedMessage id="ui-quick-marc.record.error.008.invalidValue" values={{ name }} />,
+    message: (name) => ({
+      id: 'ui-quick-marc.record.error.008.invalidValue',
+      values: { name },
+    }),
   },
   {
     tag: '010',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.multiple' }),
   },
   {
     tag: '245',
     validator: RULES.EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.title.empty" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.title.empty' }),
   },
   {
     tag: '245',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.title.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.title.multiple' }),
   },
   {
     validator: RULES.$9IN_LINKABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.$9" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.$9' }),
   },
   {
     validator: RULES.CONTROLLED_SUBFIELD,
     message: (uniqueTags) => {
       if (uniqueTags.length === 1) {
-        return (
-          <FormattedMessage
-            id="ui-quick-marc.record.error.fieldIsControlled"
-            values={{
-              count: 1,
-              fieldTags: `MARC ${uniqueTags[0]}`,
-            }}
-          />
-        );
+        return ({
+          id: 'ui-quick-marc.record.error.fieldIsControlled',
+          values: {
+            count: 1,
+            fieldTags: `MARC ${uniqueTags[0]}`,
+          },
+        });
       }
 
-      return (
-        <FormattedMessage
-          id="ui-quick-marc.record.error.fieldsAreControlled"
-          values={{
-            count: uniqueTags.length,
-            fieldTags: uniqueTags.slice(0, -1).map(tag => `MARC ${tag}`).join(', '),
-            lastFieldTag: `MARC ${uniqueTags[uniqueTags.length - 1]}`,
-          }}
-        />
-      );
+      return ({
+        id: 'ui-quick-marc.record.error.fieldsAreControlled',
+        values: {
+          count: uniqueTags.length,
+          fieldTags: uniqueTags.slice(0, -1).map(tag => `MARC ${tag}`).join(', '),
+          lastFieldTag: `MARC ${uniqueTags[uniqueTags.length - 1]}`,
+        },
+      });
     },
   },
 ];
@@ -170,28 +166,31 @@ const BASE_HOLDINGS_VALIDATORS = [
   {
     tag: '004',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.instanceHrid.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.instanceHrid.multiple' }),
   },
   {
     tag: '852',
     validator: RULES.EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.location.empty" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.location.empty' }),
   },
   {
     tag: '852',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.location.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.location.multiple' }),
   },
   {
     tag: '852',
     subfield: '$b',
     validator: RULES.NON_REPEATABLE_SUBFIELD,
-    message: (fieldTag, subField) => <FormattedMessage id="ui-quick-marc.record.error.field.onlyOneSubfield" values={{ fieldTag, subField }} />,
+    message: (fieldTag, subField) => ({
+      id: 'ui-quick-marc.record.error.field.onlyOneSubfield',
+      values: { fieldTag, subField },
+    }),
   },
   {
     tag: '852',
     validator: RULES.VALID_LOCATION,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.location.invalid" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.location.invalid' }),
   },
 ];
 
@@ -204,7 +203,7 @@ const BASE_AUTHORITY_VALIDATORS = [
       return !is010LinkedToBibRecord(initialValues.records, naturalId, linksCount);
     },
     validator: RULES.SUBFIELD_VALUE_EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.$aRemoved" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.$aRemoved' }),
   },
   {
     tag: '010',
@@ -212,34 +211,37 @@ const BASE_AUTHORITY_VALIDATORS = [
       return !is010LinkedToBibRecord(initialValues.records, naturalId, linksCount);
     },
     validator: RULES.EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.removed" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.removed' }),
   },
   {
     tag: '010',
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.multiple' }),
   },
   {
     tag: '010',
     subfield: '$a',
     validator: RULES.NON_REPEATABLE_SUBFIELD,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.$aOnlyOne" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.$aOnlyOne' }),
   },
   {
     tag: new RegExp(`${CORRESPONDING_HEADING_TYPE_TAGS.join('|')}`),
     validator: RULES.EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.heading.empty" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.heading.empty' }),
   },
   {
     tag: new RegExp(/1\d\d/),
     validator: RULES.NON_REPEATABLE,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.heading.multiple" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.heading.multiple' }),
   },
   {
     tag: new RegExp(`${CORRESPONDING_HEADING_TYPE_TAGS.join('|')}`),
     validator: RULES.TAG_CHANGED,
     ignore: ({ linksCount }) => !linksCount,
-    message: (initialTag) => <FormattedMessage id="ui-quick-marc.record.error.1xx.change" values={{ tag: initialTag }} />,
+    message: (initialTag) => ({
+      id: 'ui-quick-marc.record.error.1xx.change',
+      values: { tag: initialTag },
+    }),
   },
   {
     tag: new RegExp(`${CORRESPONDING_HEADING_TYPE_TAGS.join('|')}`),
@@ -248,11 +250,17 @@ const BASE_AUTHORITY_VALIDATORS = [
     validator: RULES.SUBFIELD_CHANGED,
     message: (changes, initialTag) => {
       if (changes.added) {
-        return <FormattedMessage id="ui-quick-marc.record.error.1xx.add$t" values={{ tag: initialTag }} />;
+        return {
+          id: 'ui-quick-marc.record.error.1xx.add$t',
+          values: { tag: initialTag },
+        };
       }
 
       if (changes.removed) {
-        return <FormattedMessage id="ui-quick-marc.record.error.1xx.remove$t" values={{ tag: initialTag }} />;
+        return {
+          id: 'ui-quick-marc.record.error.1xx.remove$t',
+          values: { tag: initialTag },
+        };
       }
 
       return null;
@@ -266,7 +274,7 @@ const CREATE_AUTHORITY_VALIDATORS = [
     tag: '010',
     validator: RULES.EXISTS,
     ignore: isFolioSourceFileNotSelected,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.absent" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.absent' }),
   },
   {
     tag: '010',
@@ -274,7 +282,7 @@ const CREATE_AUTHORITY_VALIDATORS = [
     ignore: isFolioSourceFileNotSelected,
     pattern: () => /^[a-zA-Z]/,
     validator: RULES.SUBFIELD_VALUE_MATCH,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.prefix.absent" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.prefix.absent' }),
   },
   {
     tag: '010',
@@ -286,12 +294,12 @@ const CREATE_AUTHORITY_VALIDATORS = [
       return new RegExp(`^(${codes.join('|')})([^a-zA-Z].*|$)`);
     },
     validator: RULES.SUBFIELD_VALUE_MATCH,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.010.prefix.invalid" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.010.prefix.invalid' }),
   },
   {
     tag: '001',
     validator: RULES.CONTENT_EXISTS,
-    message: () => <FormattedMessage id="ui-quick-marc.record.error.controlField.content.empty" />,
+    message: () => ({ id: 'ui-quick-marc.record.error.controlField.content.empty' }),
   },
 ];