+++ src/hooks/useValidation/useValidation.js
@@ -2,8 +2,14 @@ import {
   useCallback,
   useContext,
 } from 'react';
+import assignWith from 'lodash/assignWith';
+
 import { validators } from './rules';
 import { QuickMarcContext } from '../../contexts';
+
+const joinErrors = (errorsA, errorsB) => {
+  return assignWith({}, errorsA, errorsB, (objValue = [], srcValue = []) => objValue.concat(srcValue));
+};
 
 const useValidation = (context) => {
   const quickMarcContext = useContext(QuickMarcContext);
@@ -11,21 +17,14 @@ const useValidation = (context) => {
   const validate = useCallback((marcRecords) => {
     const validationRules = validators[context.marcType][context.action];
 
-    const errorMessage = validationRules.reduce((_errorMessage, rule) => {
-      if (_errorMessage) {
-        return _errorMessage;
-      }
+    const errors = validationRules.reduce((joinedErrors, rule) => {
+      // returns undefined or { [field.id]: 'error message' }
+      const ruleErrors = rule.validator({ ...context, ...quickMarcContext, marcRecords }, rule);
 
-      const error = rule.validator({ ...context, ...quickMarcContext, marcRecords }, rule);
+      return joinErrors(joinedErrors, ruleErrors);
+    }, {});
 
-      if (error) {
-        return error;
-      }
-
-      return undefined;
-    }, null);
-
-    return errorMessage;
+    return errors;
   }, [context, quickMarcContext]);
 
   return { validate };